Changes to gserver: - Fixed an issue with npc-controls acting as players. They will no-longer show up in playerlist or interfere with warpto commands. - Fixed an issue with npcs > id 1000 being deleted from gmaps. Also figured out how graal assigns npc ids which will fix our gmap npc id issue (not implemented yet). Level npc ids all start with id 300 and are unique per level. Database npc ids start with 1000 although I believe you can use whatever you like. Gmap npc ids start with id 100,000 (at least putnpc2 npcs do, and > 1k npc ids do delete from levels properly) - Figured out some npc properties (NPC PROP 49-52)
Changes to npc-server:
- Fixed a rounding issue when adjusting npc's position
- Database npc improvements. Sends to all NC's on login, as well as the ability to delete npcs.

+++ server/src/script/V8NPCImpl.cpp
@@ -31,7 +31,7 @@ void NPC_SetNum_X(v8::Local<v8::String> prop, v8::Local<v8::Value> value, const 
 	TNPC *npcObject = UnwrapObject<TNPC>(self);
 
 	double newValue = value->NumberValue(info.GetIsolate()->GetCurrentContext()).ToChecked();
-	int newValueInt = 16 * (int)newValue;
+	int newValueInt = 16 * newValue;
 	if (newValueInt < 0) {
 		newValueInt = (-newValueInt << 1) | 0x0001;
 	}
@@ -52,7 +52,7 @@ void NPC_SetNum_Y(v8::Local<v8::String> prop, v8::Local<v8::Value> value, const 
 	TNPC *npcObject = UnwrapObject<TNPC>(self);
 
 	double newValue = value->NumberValue(info.GetIsolate()->GetCurrentContext()).ToChecked();
-	int newValueInt = 16 * (int)newValue;
+	int newValueInt = 16 * newValue;
 	if (newValueInt < 0) {
 		newValueInt = (-newValueInt << 1) | 0x0001;
 	}
